scrapy-template/pyproject.toml
[project]
name = "scrapy-project"
version = "0.1.0"
description = "Web scraping project using Scrapy framework"
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "scrapy>=2.11.0",
    "scrapy-playwright>=0.0.34",
    "scrapy-fake-useragent>=1.4.4",
    "scrapy-rotating-proxies>=0.6.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pymongo>=4.5.0",
    "python-dotenv>=1.0.0",
    "itemloaders>=1.1.0",
    "w3lib>=2.1.2",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

monitoring = [
    "scrapy-sentry>=0.11.0",
    "scrapy-prometheus>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
  | \.pytest_cache
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["myproject"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=myproject --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"
